+++ internal/domains/auth/model/dto/dto.go
@@ -48,7 +48,7 @@ type UpdateLastLoginRequest struct {
 
 type LoginResponse struct {
 	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
+	RefreshToken string `json:"refresh_token,omitempty"`
 }
 
 func (l *LoginResponse) FromTokenPair(tokenPair *jwt.TokenPair) {
@@ -62,7 +62,7 @@ type RefreshTokenRequest struct {
 
 type RefreshTokenResponse struct {
 	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
+	RefreshToken string `json:"refresh_token,omitempty"`
 }
 
 func (r *RefreshTokenResponse) FromTokenPair(tokenPair *jwt.TokenPair) {